requirements.txt
tensorflow
numpy
opencv-python
Pillow
chess
keras